requirements.txt
Flask==2.0.3
Werkzeug==2.0.3
flask-cors==3.0.10
gunicorn==20.1.0
python-dotenv==1.0.0
requests==2.31.0
pydub==0.25.1
google-cloud-speech==2.26.0
azure-cognitiveservices-speech==1.37.0
openmeteo-requests==1.2.0
requests-cache==1.2.0
retry-requests==2.0.0
numpy==1.26.4
pandas==2.2.2
google-cloud-language==2.13.0
requests==2.31.0
urllib3==2.2.2
beautifulsoup4==4.12.3
lxml==5.2.1